requirements.txt
pyairtable>=2.3.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0

